moved the signup code to signup form in components
