--- client/src/pages/SignUpPage.jsx
@@ -1,57 +1,0 @@
-import React, { useState } from 'react';
-
-function SignupForm({ onSignup }) {
-  const [formData, setFormData] = useState({
-    name: '',
-    email: '',
-    password: ''
-  });
-
-  const handleChange = (e) => {
-    setFormData(prev => ({
-      ...prev,
-      [e.target.name]: e.target.value
-    }));
-  };
-
-  const handleSubmit = (e) => {
-    e.preventDefault();
-    onSignup(formData); // this will help send data to the backend using usestate
-  };
-
-  return (
-    <form onSubmit={handleSubmit}>
-      <h2>Sign Up</h2>
-
-      <input
-        name="name"
-        placeholder="Full Name"
-        value={formData.name}
-        onChange={handleChange}
-        required
-      />
-
-      <input
-        type="email"
-        name="email"
-        placeholder="Email"
-        value={formData.email}
-        onChange={handleChange}
-        required
-      />
-
-      <input
-        type="password"
-        name="password"
-        placeholder="Password"
-        value={formData.password}
-        onChange={handleChange}
-        required
-      />
-
-      <button type="submit">Sign Up</button>
-    </form>
-  );
-}
-
-export default SignupForm;
